requirements.txt
Flask==3.1.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
python-dotenv==1.2.1
email-validator
blinker